autogen-rag-server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
coloredlogs==15.0.1
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
executing==2.0.1
fastapi==0.108.0
filelock==3.13.1
FLAML==2.1.1
Flask==3.0.1
Flask-Cors==4.0.0
flatbuffers==23.5.26
fsspec==2023.12.2
google-auth==2.26.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
kubernetes==28.1.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.6.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
posthog==3.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyautogen==0.2.4
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
